+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Graph#nodes allows filtering by properties
+
 ## [0.1.0] - 2021-04-08
 
 - Initial release